backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.3
sqlalchemy==2.0.28
alembic==1.13.1
pymysql==1.1.0
cryptography==42.0.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==8.1.1
httpx==0.27.0
email-validator==2.1.1
